requirements.txt
pytesseract
Pillow
numpy
opencv-python
mss
pyautogui 